requirement.txt
pydantic-ai 
openai-agents 
nest-asyncio